Step one of the extraction work
